% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arctic}
\alias{arctic}
\title{Arctic Lake Sediments Compositions}
\format{
A \code{\link{data.frame}} with 4 variables:
\describe{
\item{sand}{}
\item{silt}{}
\item{clay}{}
\item{depth}{Water depth (m).}
}
}
\source{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall. \doi{10.1007/978-94-009-4109-0}.
}
\usage{
arctic
}
\description{
Sand, silt, clay compositions of 39 sediment samples at different water
depths in an Arctic lake.
}
\seealso{
Other datasets: 
\code{\link{boxite}},
\code{\link{lava}}
}
\concept{datasets}
\keyword{datasets}
