% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boxite}
\alias{boxite}
\title{Boxite Compositions}
\format{
A \code{\link{data.frame}} with 5 variables:
\describe{
\item{A}{albite.}
\item{B}{blandite.}
\item{C}{cornite.}
\item{D}{daubite.}
\item{E}{endite.}
}
}
\source{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall. \doi{10.1007/978-94-009-4109-0}.
}
\usage{
boxite
}
\description{
Compositions of 25 specimens of boxite.
}
\seealso{
Other datasets: 
\code{\link{arctic}},
\code{\link{lava}}
}
\concept{datasets}
\keyword{datasets}
