% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isopleuros-package.R
\docType{package}
\name{isopleuros-package}
\alias{isopleuros-package}
\alias{isopleuros}
\title{isopleuros: Ternary Plots}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Ternary plots made simple. This package allows to create ternary plots using 'graphics'. It provides functions to display the data in the ternary space, to add or tune graphical elements and to display statistical summaries. It also includes common ternary diagrams which are useful for the archaeologist (e.g. soil texture charts, ceramic phase diagram).
}
\details{
\tabular{ll}{
\strong{Package:} \tab isopleuros \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 1.2.0 \cr
\strong{License:} \tab GPL-3 \cr
\strong{Zenodo:} \tab \doi{10.5281/zenodo.7940389} \cr
}
}
\section{Package options}{

\code{isopleuros} uses the following \code{\link[=options]{options()}} to configure behavior:
\itemize{
\item \code{isopleuros.missing}: a \code{\link{logical}} scalar. Should \link[=NA]{missing values}
be replaced with zeros before the ternary coordinates computation
proceeds? If \code{FALSE} (the default), incomplete cases are removed.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://packages.tesselle.org/isopleuros/}
  \item \url{https://github.com/tesselle/isopleuros}
  \item Report bugs at \url{https://github.com/tesselle/isopleuros/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
33607 Pessac cedex\cr
France
}
\keyword{internal}
