% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lava}
\alias{lava}
\title{Skye Lavas Compositions}
\format{
A \code{\link{data.frame}} with 3 variables:
\describe{
\item{A}{Na2O + K2O (percent).}
\item{F}{Fe2O3 (percent).}
\item{M}{MgO (percent).}
}
}
\source{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
London: Chapman and Hall. \doi{10.1007/978-94-009-4109-0}.
}
\usage{
lava
}
\description{
AFM compositions of 23 aphyric Skye lavas.
}
\concept{datasets}
\keyword{datasets}
