% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{ternary_mean}
\alias{ternary_mean}
\alias{ternary_mean-method}
\alias{ternary_mean,numeric,numeric,numeric-method}
\alias{ternary_mean,ANY,missing,missing-method}
\title{Compositional Mean}
\usage{
ternary_mean(x, y, z, ...)

\S4method{ternary_mean}{numeric,numeric,numeric}(x, y, z, ...)

\S4method{ternary_mean}{ANY,missing,missing}(x, y, z, ...)
}
\arguments{
\item{x, y, z}{A \code{\link{numeric}} vector giving the x, y and z ternary coordinates
of a set of points. If \code{y} and \code{z} are missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xyz.coords]{grDevices::xyz.coords()}}).}

\item{...}{Further arguments to be passed to \code{\link[graphics:points]{graphics::points()}}.}
}
\value{
\code{ternary_mean()} is called it for its side-effects.
}
\description{
Computes and draws the closed geometric mean of the set of points specified.
}
\examples{
## Compositional data
coda <- data.frame(
  X = c(96, 49, 57, 92, 62, 85, 75, 76, 30, 36,
        55, 93, 78, 10, 50, 69, 50, 91, 13, 24),
  Y = c(59, 90, 18, 99, 80, 22, 74, 43, 68, 80,
        30, 84, 84, 48, 79, 53, 32, 91, 19, 48),
  Z = c(33, 88, 7, 58, 81, 16, 48, 91, 85, 12,
        94, 81, 32, 39, 39, 32, 65, 35, 4, 12)
)

## Mean
ternary_plot(coda, panel.first = ternary_grid())
ternary_mean(coda, pch = 16, col = "red")
ternary_confidence(coda, level = 0.95, border = "red", lty = 1)
}
\seealso{
Other statistics: 
\code{\link{ternary_density}()},
\code{\link{ternary_ellipse}()},
\code{\link{ternary_hull}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
