% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{ternary_points}
\alias{ternary_points}
\alias{ternary_points,numeric,numeric,numeric-method}
\alias{ternary_points,ANY,missing,missing-method}
\title{Add Points to a Ternary Plot}
\usage{
ternary_points(x, y, z, ...)

\S4method{ternary_points}{numeric,numeric,numeric}(x, y, z, type = "l", ...)

\S4method{ternary_points}{ANY,missing,missing}(x, type = "l", ...)
}
\arguments{
\item{x, y, z}{A \code{\link{numeric}} vector giving the x, y and z ternary coordinates
of a set of points. If \code{y} and \code{z} are missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xyz.coords]{grDevices::xyz.coords()}}).}

\item{...}{Further graphical parameters (see \code{\link[graphics:par]{graphics::par()}}) may also be
supplied as arguments, particularly, plotting character, \code{pch}, character
expansion, \code{cex} and color, \code{col}.}

\item{type}{A \code{\link{character}} string indicating the type of plotting; actually
any of the types as in \code{\link[graphics:plot.default]{graphics::plot.default()}}.}
}
\value{
\code{ternary_points()} is called it for its side-effects.
}
\description{
Add Points to a Ternary Plot
}
\examples{
## Compositional data
coda <- data.frame(
  X = c(1, 0, 0, 1/3),
  Y = c(0, 1, 0, 1/3),
  Z = c(0, 0, 1, 1/3)
)

## Ternary plot
ternary_plot(coda)
ternary_plot(coda, pch = 16, col = "red", cex = 1:4)

## Add a grid
ternary_grid(5, 0)
ternary_plot(coda, panel.first = ternary_grid(5, 10))

## Add axis
ternary_plot(coda, axes = FALSE)
ternary_axis(side = 1, col = "red")
ternary_axis(side = 2, col = "blue")
ternary_axis(side = 3, col = "green")

## Zoom
ternary_plot(coda, xlim = c(0.5, 1), panel.first = ternary_grid())
ternary_plot(coda, ylim = c(0.5, 1), panel.first = ternary_grid())
ternary_plot(coda, zlim = c(0.5, 1), panel.first = ternary_grid())
}
\seealso{
\code{\link[graphics:points]{graphics::points()}}

Other plot methods: 
\code{\link{ternary_axis}()},
\code{\link{ternary_grid}()},
\code{\link{ternary_lines}()},
\code{\link{ternary_plot}()},
\code{\link{ternary_polygon}()}
}
\author{
N. Frerebeau
}
\concept{plot methods}
