% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{ternary_polygon}
\alias{ternary_polygon}
\alias{ternary_polygon,numeric,numeric,numeric-method}
\alias{ternary_polygon,ANY,missing,missing-method}
\title{Polygon Drawing}
\usage{
ternary_polygon(x, y, z, ...)

\S4method{ternary_polygon}{numeric,numeric,numeric}(x, y, z, ...)

\S4method{ternary_polygon}{ANY,missing,missing}(x, y, z, ...)
}
\arguments{
\item{x, y, z}{A \code{\link{numeric}} vector giving the x, y and z ternary coordinates
of a set of points. If \code{y} and \code{z} are missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xyz.coords]{grDevices::xyz.coords()}}).}

\item{...}{Further arguments to be passed to \code{\link[graphics:polygon]{graphics::polygon()}}.}
}
\value{
\code{ternary_polygon()} is called it for its side-effects.
}
\description{
Draws the polygons whose vertices are given in \code{x}, \code{y} and \code{z}.
}
\examples{
## Compositional data
coda <- data.frame(
  X = c(20, 60, 20, 20),
  Y = c(20, 20, 60, 40),
  Z = c(60, 20, 20, 40)
)

## Ternary plot
ternary_plot(coda, panel.first = ternary_grid(5, 10))

## Add points
ternary_plot(NULL, panel.first = ternary_grid(5, 10))
ternary_points(coda, col = "red", pch = 16)

## Add lines
ternary_plot(NULL, panel.first = ternary_grid(5, 10))
ternary_lines(coda, col = "red", lwd = 2)

## Add a polygon
ternary_plot(NULL, panel.first = ternary_grid(5, 10))
ternary_polygon(coda, density = 5, border = "red")

## Add text
ternary_plot(NULL, panel.first = ternary_grid(5, 10))
ternary_text(coda, col = "red")

## Add arrows
ternary_plot(NULL, panel.first = ternary_grid(5, 10))
ternary_arrows(x0 = 40, y0 = 20, z0 = 40,
               x1 = 20, y1 = 40, z1 = 40)
}
\seealso{
\code{\link[graphics:polygon]{graphics::polygon()}}

Other plot methods: 
\code{\link{ternary_arrows}()},
\code{\link{ternary_axis}()},
\code{\link{ternary_grid}()},
\code{\link{ternary_lines}()},
\code{\link{ternary_plot}()},
\code{\link{ternary_points}()},
\code{\link{ternary_text}()}
}
\author{
N. Frerebeau
}
\concept{plot methods}
