% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coordinates.R
\docType{methods}
\name{coordinates_ternary}
\alias{coordinates_ternary}
\alias{coordinates_ternary-method}
\alias{coordinates_ternary,numeric,numeric,numeric-method}
\alias{coordinates_ternary,ANY,missing,missing-method}
\title{Ternary Coordinates}
\usage{
coordinates_ternary(x, y, z, ...)

\S4method{coordinates_ternary}{numeric,numeric,numeric}(
  x,
  y,
  z,
  center = FALSE,
  scale = FALSE,
  missing = getOption("isopleuros.missing")
)

\S4method{coordinates_ternary}{ANY,missing,missing}(
  x,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  center = FALSE,
  scale = FALSE,
  missing = getOption("isopleuros.missing")
)
}
\arguments{
\item{x, y, z}{A \code{\link{numeric}} vector giving the x, y and z Cartesian coordinates
of a set of points. If \code{y} and \code{z} are missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xyz.coords]{grDevices::xyz.coords()}}).}

\item{...}{Currently not used.}

\item{center}{A \code{\link{logical}} scalar or a \code{\link{numeric}} vector giving the center.}

\item{scale}{A \code{\link{logical}} scalar or a length-one \code{\link{numeric}} vector giving a
scaling factor.}

\item{missing}{A \code{\link{logical}} scalar: should \link[=NA]{missing values} be replaced
with zeros before the computation proceeds? If \code{FALSE} (the default),
incomplete cases are removed.}

\item{xlab, ylab, zlab}{A \code{\link{character}} string specifying the names for the x,
y and z variables to be extracted.}
}
\value{
A \code{\link{list}} with the components:
\describe{
\item{\code{x}}{A \code{\link{numeric}} vector of x coordinates.}
\item{\code{y}}{A \code{\link{numeric}} vector of y coordinates.}
\item{\code{center}}{A \code{\link{numeric}} vector giving the center.}
\item{\code{scale}}{A \code{\link{numeric}} vector giving the scale factor.}
}
}
\description{
Computes ternary coordinates.
}
\examples{
## Compositional data
coda <- data.frame(
  x = c(1, 0, 0, 1/3),
  y = c(0, 1, 0, 1/3),
  z = c(0, 0, 1, 1/3)
)

## Ternary coordinates
(tern <- coordinates_ternary(coda))

## Cartesian coordinates
(cart <- coordinates_cartesian(tern))
}
\seealso{
Other coordinates: 
\code{\link{coordinates_cartesian}()}
}
\author{
N. Frerebeau
}
\concept{coordinates}
\keyword{internal}
