% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{ternary_density}
\alias{ternary_density}
\alias{ternary_density-method}
\alias{ternary_density,numeric,numeric,numeric-method}
\alias{ternary_density,ANY,missing,missing-method}
\title{Density Contour Lines}
\usage{
ternary_density(x, y, z, ...)

\S4method{ternary_density}{numeric,numeric,numeric}(
  x,
  y,
  z,
  h = NULL,
  n = 25,
  nlevels = 10,
  levels = NULL,
  col = graphics::par("fg"),
  lty = graphics::par("lty"),
  lwd = graphics::par("lwd"),
  ...
)

\S4method{ternary_density}{ANY,missing,missing}(
  x,
  h = NULL,
  n = 25,
  nlevels = 10,
  levels = NULL,
  col = graphics::par("fg"),
  lty = graphics::par("lty"),
  lwd = graphics::par("lwd"),
  ...
)
}
\arguments{
\item{x, y, z}{A \code{\link{numeric}} vector giving the x, y and z ternary coordinates
of a set of points. If \code{y} and \code{z} are missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xyz.coords]{grDevices::xyz.coords()}}).}

\item{...}{Further arguments to be passed to \code{\link[=ternary_lines]{ternary_lines()}}.}

\item{h}{A length-one \code{\link{numeric}} vector giving the bandwidth.}

\item{n}{A length-one \code{\link{numeric}} specifying the number of grid points.}

\item{nlevels}{A length-one \code{\link{numeric}} vector specifying the number of
contour levels desired. Only used if \code{levels} is \code{NULL}.}

\item{levels}{A \code{\link{numeric}} vector of levels at which to draw contour lines.}

\item{col}{Colors for the lines. If there are fewer colours than \code{levels},
the colours will be interpolated and mapped to the level values.}

\item{lty}{A \code{\link{character}} string or \code{\link{numeric}} value specifying the line
types.}

\item{lwd}{A non-negative \code{\link{numeric}} value specifying the line widths.}
}
\value{
\code{ternary_density()} is called it for its side-effects.
}
\description{
Computes and draws density contour lines.
}
\details{
Two-dimensional kernel density estimation with an axis-aligned bivariate
normal kernel. Normal kernel is evaluated on a square grid, after an
isometric log ratio transformation of the original data.
}
\note{
Two-dimensional kernel density estimation is adapted from \code{\link[MASS:kde2d]{MASS::kde2d()}}.

This must be considered as experimental and subject to major changes
in a future release.
}
\examples{
## Data from Aitchison 1986
ternary_plot(lava, panel.first = ternary_grid())

## Add density
ternary_density(lava, n = 500, nlevels = 10, col = c("yellow", "red"))
}
\seealso{
\code{\link[=ternary_lines]{ternary_lines()}}

Other statistics: 
\code{\link{ternary_ellipse}()},
\code{\link{ternary_hull}()},
\code{\link{ternary_mean}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
