% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{ternary_ellipse}
\alias{ternary_ellipse}
\alias{ternary_confidence}
\alias{ternary_tolerance}
\alias{ternary_ellipse,numeric,numeric,numeric-method}
\alias{ternary_ellipse,ANY,missing,missing-method}
\alias{ternary_confidence,numeric,numeric,numeric-method}
\alias{ternary_confidence,ANY,missing,missing-method}
\alias{ternary_tolerance,numeric,numeric,numeric-method}
\alias{ternary_tolerance,ANY,missing,missing-method}
\title{Add an Ellipse to a Ternary Plot}
\usage{
ternary_ellipse(x, y, z, ...)

ternary_confidence(x, y, z, ...)

ternary_tolerance(x, y, z, ...)

\S4method{ternary_ellipse}{numeric,numeric,numeric}(x, y, z, radius = 1, ...)

\S4method{ternary_ellipse}{ANY,missing,missing}(x, radius = 1, ...)

\S4method{ternary_confidence}{numeric,numeric,numeric}(x, y, z, level = 0.95, ...)

\S4method{ternary_confidence}{ANY,missing,missing}(x, level = 0.95, ...)

\S4method{ternary_tolerance}{numeric,numeric,numeric}(x, y, z, level = 0.95, ...)

\S4method{ternary_tolerance}{ANY,missing,missing}(x, level = 0.95, ...)
}
\arguments{
\item{x, y, z}{A \code{\link{numeric}} vector giving the x, y and z ternary coordinates
of a set of points. If \code{y} and \code{z} are missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xyz.coords]{grDevices::xyz.coords()}}).}

\item{...}{Further arguments to be passed to \code{\link[graphics:polygon]{graphics::polygon()}}.}

\item{radius}{A \code{\link{numeric}} vector specifying the scaling of the
half-diameters.}

\item{level}{A \code{\link{numeric}} vector specifying the confidence/tolerance level.}
}
\value{
\code{ternary_ellipse()} is called it for its side-effects.
}
\description{
Computes and draws a confidence/tolerance ellipse.
}
\examples{
## Compositional data
coda <- data.frame(
  X = c(96, 49, 57, 92, 62, 85, 75, 76, 30, 36,
        55, 93, 78, 10, 50, 69, 50, 91, 13, 24),
  Y = c(59, 90, 18, 99, 80, 22, 74, 43, 68, 80,
        30, 84, 84, 48, 79, 53, 32, 91, 19, 48),
  Z = c(33, 88, 7, 58, 81, 16, 48, 91, 85, 12,
        94, 81, 32, 39, 39, 32, 65, 35, 4, 12)
)

## Ellipse
ternary_plot(coda, panel.first = ternary_grid(5, 10))
ternary_tolerance(coda, level = 0.95, border = "blue", lty = 2)
ternary_confidence(coda, level = 0.95, border = "red", lty = 3)
}
\seealso{
\code{\link[graphics:polygon]{graphics::polygon()}}

Other statistics: 
\code{\link{ternary_hull}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
