% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{ternary_hull}
\alias{ternary_hull}
\alias{ternary_hull-method}
\alias{ternary_hull,numeric,numeric,numeric-method}
\alias{ternary_hull,ANY,missing,missing-method}
\title{Convex Hull of a Set of Points}
\usage{
ternary_hull(x, y, z, ...)

\S4method{ternary_hull}{numeric,numeric,numeric}(x, y, z, ...)

\S4method{ternary_hull}{ANY,missing,missing}(x, y, z, ...)
}
\arguments{
\item{x, y, z}{A \code{\link{numeric}} vector giving the x, y and z ternary coordinates
of a set of points. If \code{y} and \code{z} are missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xyz.coords]{grDevices::xyz.coords()}}).}

\item{...}{Further arguments to be passed to \code{\link[graphics:polygon]{graphics::polygon()}}.}
}
\value{
\code{ternary_hull()} is called it for its side-effects.
}
\description{
Computes and draws the convex hull of the set of points specified.
}
\examples{
## Compositional data
coda <- data.frame(
  X = c(96, 49, 57, 92, 62, 85, 75, 76, 30, 36,
        55, 93, 78, 10, 50, 69, 50, 91, 13, 24),
  Y = c(59, 90, 18, 99, 80, 22, 74, 43, 68, 80,
        30, 84, 84, 48, 79, 53, 32, 91, 19, 48),
  Z = c(33, 88, 7, 58, 81, 16, 48, 91, 85, 12,
        94, 81, 32, 39, 39, 32, 65, 35, 4, 12)
)

## Convex hull
ternary_plot(coda, panel.first = ternary_grid(5, 10))
ternary_hull(coda, border = "red")
}
\seealso{
\code{\link[grDevices:chull]{grDevices::chull()}}, \code{\link[graphics:polygon]{graphics::polygon()}}

Other statistics: 
\code{\link{ternary_ellipse}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
